arrange paths
